+++ src/main/java/kr/santanrudolph/everyvent/domain/calendar/service/CalendarService.java
@@ -1,15 +1,11 @@
 package kr.santanrudolph.everyvent.domain.calendar.service;
 
 import kr.santanrudolph.everyvent.domain.calendar.Calendar;
-import kr.santanrudolph.everyvent.domain.calendar.CalendarConstants;
 import kr.santanrudolph.everyvent.domain.calendar.dto.*;
 import kr.santanrudolph.everyvent.domain.calendar.enums.CalendarType;
-import kr.santanrudolph.everyvent.domain.calendar.enums.Visibility;
 import kr.santanrudolph.everyvent.domain.calendar.repository.CalendarRepository;
-import kr.santanrudolph.everyvent.domain.follow.FollowService;
 import kr.santanrudolph.everyvent.domain.user.User;
 import kr.santanrudolph.everyvent.domain.user.UserService;
-import kr.santanrudolph.everyvent.domain.user.enums.Role;
 import kr.santanrudolph.everyvent.global.exception.ErrorCode;
 import kr.santanrudolph.everyvent.global.exception.EveryventException;
 import lombok.RequiredArgsConstructor;
@@ -36,7 +32,7 @@ public class CalendarService {
     private final CalendarRepository calendarRepository;
     private final UserService userService;
     private final ScrapService scrapService;
-    private final FollowService followService;
+    private final CalendarPolicy calendarPolicy;
 
 
     @Transactional
@@ -44,12 +40,7 @@ public class CalendarService {
 
         User user = userService.getCurrentUser();
 
-        try {
-            validateCalendarLimit(user, request.startDate());
-            validateFuture(request.startDate());
-        } catch (EveryventException e) {
-            throw new EveryventException(ErrorCode.INVALID_INPUT, "캘린더를 만들 수 없습니다. " + e.getMessage());
-        }
+        calendarPolicy.validateCanCreate(user, request.startDate());
 
         Calendar calendar = Calendar.createCalendarWithStartDay(
                 user,
@@ -74,10 +65,10 @@ public class CalendarService {
         Calendar calendar = findCalendarByIdAndDeletedAtIsNull(calendarId);
         User currentUser = userService.getCurrentUser();
 
-        validateCanViewCalendar(currentUser, calendar);
+        calendarPolicy.validateCanView(currentUser, calendar);
 
         Long scrapCount = scrapService.getScrapCount(calendarId);
-        boolean scrappable = isScrappable(currentUser, calendar);
+        boolean scrappable = calendarPolicy.canScrap(currentUser, calendar);
         return CalendarDetailResponse.from(calendar, scrappable, scrapCount);
     }
 
@@ -154,7 +145,7 @@ public class CalendarService {
         User currentUser = userService.getCurrentUser();
         Calendar calendar = findCalendarByIdAndDeletedAtIsNull(calendarId);
 
-        validateUpdateCalendar(currentUser, calendar);
+        calendarPolicy.validateCanUpdate(currentUser, calendar);
 
         if (request.title() != null) {
             calendar.updateTitle(request.title());
@@ -195,7 +186,7 @@ public class CalendarService {
         Calendar calendar = calendarRepository.findById(calendarId)
                 .orElseThrow(() -> new EveryventException(ErrorCode.NOT_FOUND, "캘린더를 찾을 수 없습니다."));
 
-        validateUpdateScrapColor(currentUser, calendar);
+        calendarPolicy.validateCanUpdateScrapColor(currentUser, calendar);
 
         calendar.updateColor(request.color());
 
@@ -203,16 +194,20 @@ public class CalendarService {
         return CalendarDetailResponse.from(calendar, SCRAPPABLE, scrapCount);
     }
 
-
     @Transactional
     public void deleteCalendar(Long calendarId) {
         User currentUser = userService.getCurrentUser();
 
         Calendar calendar = findCalendarByIdAndDeletedAtIsNull(calendarId);
 
-        validateCanDeleteCalendar(currentUser, calendar);
+        calendarPolicy.validateCanDelete(currentUser, calendar);
 
         calendar.softDelete();
+    }
+
+    public Calendar findCalendarByIdAndDeletedAtIsNull(Long calendarId) {
+        return calendarRepository.findByIdAndDeletedAtIsNull(calendarId)
+                .orElseThrow(() -> new EveryventException(ErrorCode.NOT_FOUND, "캘린더를 찾을 수 없습니다."));
     }
 
     private Map<Long, Long> getScrapCountMapFromCalendars(List<Calendar> calendars) {
@@ -248,139 +243,6 @@ public class CalendarService {
                         entry.getValue()
                 ))
                 .collect(Collectors.toCollection(ArrayList::new));
-    }
-
-    private Calendar findCalendarByIdAndDeletedAtIsNull(Long calendarId) {
-        return calendarRepository.findByIdAndDeletedAtIsNull(calendarId)
-                .orElseThrow(() -> new EveryventException(ErrorCode.NOT_FOUND, "캘린더를 찾을 수 없습니다."));
-    }
-
-    boolean canViewCalendar(User user, Calendar calendar) {
-        if (user.getRole() == Role.ADMIN) { // 관리자는 다 조회 가능.
-            return true;
-        }
-
-        if (isOwner(user, calendar)) {
-            return true;
-        }
-
-        Visibility visibility = calendar.getVisibility();
-        if (visibility == Visibility.PUBLIC) {
-            return true;
-        }
-
-        Long followerId = user.getId();
-        Long targetId = calendar.getUser().getId();
-        if (visibility == Visibility.FOLLOWER) {
-            return followService.isFollowing(followerId, targetId);
-        }
-
-        if (visibility == Visibility.MUTUAL) {
-            return followService.isMutualFollow(followerId, targetId);
-        }
-
-        return false;
-    }
-
-    boolean isScrappable(User user, Calendar calendar) {
-        if (isOwner(user, calendar)) {
-            return false;
-        }
-
-        if (!canViewCalendar(user, calendar)) {
-            return false;
-        }
-
-        return !scrapService.isScrapped(user, calendar.getId());
-    }
-
-    boolean isOwner(User user, Calendar calendar) {
-        return calendar.getUser().getId().equals(user.getId());
-    }
-
-    private void validateCanViewCalendar(User user, Calendar calendar) {
-        if (canViewCalendar(user, calendar)) {
-            return;
-        }
-        throw new EveryventException(ErrorCode.INVALID_INPUT, "해당 캘린더에 접근 권한이 없습니다.");
-    }
-
-    private void validateCalendarLimit(User user, LocalDate targetDate) {
-        int year = targetDate.getYear();
-        int month = targetDate.getMonthValue();
-
-        long count = calendarRepository.countByUserAndYearMonthAndTypes(
-                user, year, month, CalendarConstants.LIMITED_CALENDAR_TYPES
-        );
-
-        if (count >= CalendarConstants.MAX_CNT_PER_MONTH) {
-            throw new EveryventException(
-                    ErrorCode.INVALID_INPUT,
-                    String.format("%d년 %d월에 생성 가능한 캘린더 수를 초과했습니다.", year, month)
-            );
-        }
-    }
-
-    private void validateUpdateCalendar(User user, Calendar calendar) {
-        try {
-            validateFuture(calendar.getStartDate());
-            validateDeleted(calendar);
-            validateOwner(user, calendar);
-            validateOriginalCalendar(calendar);
-        } catch (EveryventException e) {
-            throw new EveryventException(ErrorCode.INVALID_INPUT, "캘린더를 수정할 수 없습니다. " + e.getMessage());
-        }
-    }
-
-    private void validateUpdateScrapColor(User user, Calendar calendar) {
-        try {
-            validateFuture(calendar.getStartDate());
-            validateDeleted(calendar);
-            validateOwner(user, calendar);
-            validateScrappedCalendar(calendar);
-        } catch (EveryventException e) {
-            throw new EveryventException(ErrorCode.INVALID_INPUT, "캘린더 색상을 수정할 수 없습니다. " + e.getMessage());
-        }
-    }
-
-    private void validateCanDeleteCalendar(User user, Calendar calendar) {
-        validateFuture(calendar.getStartDate());
-        validateDeleted(calendar);
-        validateOwner(user, calendar);
-    }
-
-    private void validateFuture(LocalDate startDate) {
-        YearMonth calendarYearMonth = YearMonth.from(startDate);
-        YearMonth currentYearMonth = YearMonth.now();
-
-        if (calendarYearMonth.isAfter(currentYearMonth)) {
-            return;
-        }
-        throw new EveryventException(ErrorCode.INVALID_INPUT, "미래의 캘린더가 아닙니다.");
-    }
-
-    private void validateDeleted(Calendar calendar) {
-        if (calendar.isDeleted()) {
-            throw new EveryventException(ErrorCode.NOT_FOUND, "삭제된 캘린더입니다.");
-        }
-    }
-
-    private void validateOwner(User user, Calendar calendar) {
-        if (!isOwner(user, calendar)) {
-            throw new EveryventException(ErrorCode.FORBIDDEN, "본인 캘린더가 아닙니다.");
-        }
-    }
-
-    private void validateOriginalCalendar(Calendar calendar) {
-        if (!calendar.isOriginalCalendar()) {
-            throw new EveryventException(ErrorCode.FORBIDDEN, "원본 캘린더가 아닙니다.");
-        }
-    }
-
-    private void validateScrappedCalendar(Calendar calendar) {
-        if (calendar.getCalendarType().equals(CalendarType.SCRAPED)) {
-            throw new EveryventException(ErrorCode.INVALID_INPUT, "스크랩한 캘린더가 아닙니다.");
-        }
     }
 
 }
